+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 Updated
 ~~~~~~~
 
+-  Pin black version to 20.8b1 for local dev env consistency
 -  Remove util.robotize/humanize and change callers to use autoprotocol directly
 
 v8.0.0