+++ CHANGELOG.rst
@@ -9,6 +9,13 @@ Added
 
 - Codecov configuration for coverage
 
+v8.1.1
+------
+Fixed
+~~~~~
+
+- Encoding error when using dataset upload API with request signing enabled
+- Bad request error when making calls to non-Strateos endpoints (e.g. S3) with authorization headers set
 
 v8.1.0
 ------